+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # OCaml Expect and Inline Test Explorer for Visual Studio Code Changelog
 
+## Version 0.2.0 (2023-03-06)
+
+- Remove unnecessary node 'Expect and Inline Tests' in the Test Explorer tree.
+- Change the name of the test profile to 'Run Expect and Inline PPX tests'.
+- Add documentation of 'Run Profiles'.
+- Add GitHub Issue Template.
+
 ## Version 0.1.0 (2023-03-05)
 
 Initial release